* Move classes in packages
