+++ src/test/java/com/coveo/spillway/storage/AsyncLimitUsageStorageTest.java
@@ -23,9 +23,9 @@
 package com.coveo.spillway.storage;
 
 import com.coveo.spillway.limit.LimitKey;
-import com.coveo.spillway.limit.LimitUsageStorage;
 import com.coveo.spillway.storage.AsyncLimitUsageStorage;
-import com.coveo.spillway.storage.AddAndGetRequest.Builder;
+import com.coveo.spillway.storage.utils.AddAndGetRequest;
+import com.coveo.spillway.storage.utils.AddAndGetRequest.Builder;
 import com.google.common.collect.ImmutableMap;
 
 import org.junit.Before;
